changed datatype of decimal to float for json conversion

+++ database/createtable.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS salaries;
 CREATE TABLE salaries (
     id INT NOT NULL,
     team TEXT NOT NULL,
-    winpercentage DECIMAL(4,3) NOT NULL,
+    winpercentage FLOAT NOT NULL,
     teamyear SMALLINT NOT NULL,
     roundmade SMALLINT NOT NULL,
     positiongroup TEXT NOT NULL,
@@ -14,7 +14,7 @@ CREATE TABLE salaries (
 DROP TABLE IF EXISTS salarieswide;
 CREATE TABLE salarieswide (
     team TEXT NOT NULL,
-    winpercentage DECIMAL(4,3) NOT NULL,
+    winpercentage FLOAT NOT NULL,
     catchers TEXT NOT NULL,
     infielders TEXT NOT NULL,
     outfielders TEXT NOT NULL,
